Optimizes memory use when printing all configurations
Previously, printing all configurations relied on a stream that
eventually held all of them in memory; then, for each configuration
coming from such stream, we queried the ADD to get the corresponding
result. This was not only a memory hog, but also consumed more CPU.

This new implementation leverages the fact that CUDD's cube iterator
(which is also used to enumerate valid configurations) also provides the
value (ADD result) for each cube (combination of variables).
Thus, we adapted the configuration iterator to run at once over
configurations and values, printing them all.