Remove noref ADD initializer.
I think that there could be problems from calling Cudd_ref on an ADD that has been
read by Dddmp, since that is called inside the Dddmp_cuddAddLoad already.
In order to not introduce changes, however, I decided to keep the
ADD initialization as it is in the master branch.